Move logic code from BrazeMessages to logic
