+++ logic-index.ts
@@ -1,2 +1,2 @@
-export * from './src/BrazeMessages/BrazeMessages';
-export * from './src/BrazeMessages/LocalMessageCache';
+export * from './src/logic/BrazeMessages';
+export * from './src/logic/LocalMessageCache';
